Enum roxmltree::Error

pub enum Error {

    InvalidXmlPrefixUri(TextPos),
    UnexpectedXmlUri(TextPos),
    UnexpectedXmlnsUri(TextPos),
    InvalidElementNamePrefix(TextPos),
    DuplicatedNamespace(String, TextPos),
    UnknownNamespace(String, TextPos),
    UnexpectedCloseTag {
        expected: String,
        actual: String,
        pos: TextPos,
    },
    UnexpectedEntityCloseTag(TextPos),
    UnknownEntityReference(String, TextPos),
    MalformedEntityReference(TextPos),
    EntityReferenceLoop(TextPos),
    InvalidAttributeValue(TextPos),
    DuplicatedAttribute(String, TextPos),
    NoRootNode,
    UnclosedRootNode,
    DtdDetected,
    NodesLimitReached,
    AttributesLimitReached,
    NamespacesLimitReached,
    ParserError(Error),

}

A list of all possible errors.

Variants

InvalidXmlPrefixUri(TextPos)

The xmlns:xml attribute must have an http://www.w3.org/XML/1998/namespace URI.

UnexpectedXmlUri(TextPos)

Only the xmlns:xml attribute can have the http://www.w3.org/XML/1998/namespace URI.

UnexpectedXmlnsUri(TextPos)

The http://www.w3.org/2000/xmlns/ URI must not be declared.

InvalidElementNamePrefix(TextPos)

xmlns canโt be used as an element prefix.

DuplicatedNamespace(String, TextPos)

A namespace was already defined on this element.

UnknownNamespace(String, TextPos)

An unknown namespace.

Indicates that an element or an attribute has an unknown qualified name prefix.

The first value is a prefix.

UnexpectedCloseTag

Fields

expected: String

actual: String

pos: TextPos

Incorrect tree structure.

UnexpectedEntityCloseTag(TextPos)

Entity value starts with a close tag.

Example:

<!DOCTYPE test [ <!ENTITY p '</p>'> ]>
<root>&p;</root>

UnknownEntityReference(String, TextPos)

A reference to an entity that was not defined in the DTD.

MalformedEntityReference(TextPos)

A malformed entity reference.

A & character inside an attribute value or text indicates an entity reference. Otherwise, the document is not well-formed.

EntityReferenceLoop(TextPos)

A possible entity reference loop.

The current depth limit is 10. The max number of references per reference is 255.

InvalidAttributeValue(TextPos)

Attribute value cannot have a < character.

DuplicatedAttribute(String, TextPos)

An element has a duplicated attributes.

This also includes namespaces resolving. So an element like this will lead to an error.

<e xmlns:n1='http://www.w3.org' xmlns:n2='http://www.w3.org' n1:a='b1' n2:a='b2'/>

NoRootNode

The XML document must have at least one element.

UnclosedRootNode

The root node was opened but never closed.

DtdDetected

An XML with DTD detected.

This error will be emitted only when ParsingOptions::allow_dtd is set to false.

NodesLimitReached

Indicates that the ParsingOptions::nodes_limit was reached.

AttributesLimitReached

Indicates that too many attributes were parsed.

NamespacesLimitReached

Indicates that too many namespaces were parsed.

ParserError(Error)

Errors detected by the xmlparser crate.

Implementations

impl Error

pub fn pos(&self) -> TextPos

Returns the error position.